login

+++ app/routes/customer.py
@@ -1,41 +1,43 @@
-from flask import jsonify, request
+from flask import jsonify, request, Flask, session
+from hashlib import sha256
+from base64 import b64encode, b64decode
 from app import app, db
 from app.models import Customer, Merchant, Transaction
+from app.middleware import loginMiddleware
 from sqlalchemy import text
 from sqlalchemy.exc import SQLAlchemyError
-from cryptography.hazmat.primitives import padding
-from cryptography.hazmat.primitives.ciphers import Cipher, algorithms, modes
 from cryptography.hazmat.backends import default_backend
-import base64
+from cryptography.hazmat.primitives import hashes
+from cryptography.hazmat.primitives.ciphers import Cipher, algorithms, modes
+from Crypto.Cipher import AES
+from Crypto.Util.Padding import pad, unpad
+from werkzeug.security import generate_password_hash, check_password_hash
 import os
 
-# Fungsi untuk mengenkripsi password menggunakan AES
-def encrypt_password(password):
-    key = os.urandom(16)
-    backend = default_backend()
-    iv = os.urandom(16)
-    padder = padding.PKCS7(128).padder()
-    padded_data = padder.update(password.encode()) + padder.finalize()
-    cipher = Cipher(algorithms.AES(key), modes.CBC(iv), backend=backend)
-    encryptor = cipher.encryptor()
-    ciphertext = encryptor.update(padded_data) + encryptor.finalize()
-    return base64.b64encode(ciphertext).decode()
+# key = os.urandom(16)
+key = sha256(os.urandom(32)).digest()
 
-# Fungsi untuk mendekripsi password menggunakan AES
-def decrypt_password(encrypted_password):
-    ciphertext = base64.b64decode(encrypted_password.encode())
-    backend = default_backend()
-    iv = ciphertext[:16]
-    cipher = Cipher(algorithms.AES(key), modes.CBC(iv), backend=backend)
-    decryptor = cipher.decryptor()
-    padded_data = decryptor.update(ciphertext[16:]) + decryptor.finalize()
-    unpadder = padding.PKCS7(128).unpadder()
-    password = unpadder.update(padded_data) + unpadder.finalize()
-    return password.decode()
+def encrypt_password(plaintext):
+    iv = os.urandom(16)    
+    cipher = AES.new(key, AES.MODE_CFB, iv=iv)
+    ciphertext = cipher.encrypt(pad(plaintext.encode(), AES.block_size))
+    return b64encode(iv + ciphertext).decode()
+
+def decrypt_password(ciphertext):    
+    ciphertext = b64decode(ciphertext.encode())    
+    iv = ciphertext[:16]    
+    ciphertext = ciphertext[16:]    
+    cipher = AES.new(key, AES.MODE_CFB, iv=iv)
+    plaintext = unpad(cipher.decrypt(ciphertext), AES.block_size).decode()
+    return plaintext
+
 
 # add customer
 @app.route('/customer', methods=['POST'])
 def create_customer():
+    if check_login_customer(session.get('customer'),session.get('level')) : 
+       return jsonify({'error': 'Access Forbidden'}), 403 
+
     data = request.get_json()
     username = data.get('username')
     email = data.get('email')
@@ -45,9 +47,9 @@ def create_customer():
     if not (username and email and password and no_rek):
         return jsonify({'error': 'Incomplete data provided'}), 400
     
-    encrypted_password = encrypt_password(password)
 
     try:
+        encrypted_password = encrypt_password(password)       
         query = text("INSERT INTO ms_customer (username, email, password, no_rek) VALUES (:username, :email, :password, :no_rek)")
         db.session.execute(query, {"username": username, "email": email, "password": encrypted_password, "no_rek": no_rek})
         db.session.commit()
@@ -60,12 +62,15 @@ def create_customer():
 # Get all customer
 @app.route('/customer', methods=['GET'])
 def get_customers():
+    if check_login_customer(session.get('customer'),session.get('level')) : 
+       return jsonify({'error': 'Access Forbidden'}), 403 
+
     try:
-        query = text("SELECT id, username, email, no_rek FROM ms_customer")
+        query = text("SELECT id, username, email, no_rek, password FROM ms_customer")
         result = db.session.execute(query)
         result_array = [row for row in result]
 
-        result_dicts = [{'id': row[0],'username': row[1], 'email': row[2], 'no_rek' : row[3]} for row in result_array]
+        result_dicts = [{'id': row[0],'username': row[1], 'email': row[2], 'no_rek' : row[3],'password' : row[4]} for row in result_array]
 
         if not result_dicts:            
             return jsonify([]), 200 
@@ -78,6 +83,9 @@ def get_customers():
 # Update customer
 @app.route('/customer/<int:id>', methods=['PUT'])
 def update_customer(id):
+    if check_login_customer(session.get('customer'),session.get('level')) : 
+       return jsonify({'error': 'Access Forbidden'}), 403 
+    
     data = request.get_json()
     username = data.get('username')
     email = data.get('email')
@@ -101,6 +109,9 @@ def update_customer(id):
 # Delete customer
 @app.route('/customer/<int:id>', methods=['DELETE'])
 def delete_customer(id):
+    if check_login_customer(session.get('customer'),session.get('level')) : 
+       return jsonify({'error': 'Access Forbidden'}), 403 
+
     try:
         query = text("DELETE FROM ms_customer WHERE id=:id")
         db.session.execute(query,  {'id': id})
@@ -108,3 +119,30 @@ def delete_customer(id):
         return jsonify({'message': 'Customer deleted successfully!'}), 200
     except Exception as e:
         return jsonify({'error': 'Failed to delete customer', 'details': str(e)}), 500
+
+
+@app.route('/login/customer', methods=['POST'])
+def login():
+    if check_login(session.get('customer')):
+        return jsonify({'error': 'Login Already'}), 403
+
+    data = request.get_json()
+    email = data.get('email')
+    password = data.get('password')
+
+    if not (email and password):
+        return jsonify({'error': 'Incomplete data provided'}), 400    
+
+    try:
+        customer = Customer.query.filter_by(email=email).first()
+        if customer:    
+            if password == decrypt_password(customer.password):  
+                session['customer'] = customer.username                         
+                session['level'] = "customer"
+                return jsonify({'message': 'Login successful'}), 200
+            else:                
+                return jsonify({'error': 'Incorrect password'}), 401
+        else:            
+            return jsonify({'error': 'Customer not registered'}), 404
+    except SQLAlchemyError as e:        
+        return jsonify({'error': 'Database error'}), 500
